+++ wire/message.go
@@ -35,9 +35,11 @@ const (
 	CmdAddr           = "addr"
 	CmdGetBlocks      = "getblocks"
 	CmdInv            = "inv"
+	CmdGetMissedTxs   = "getmissedtxs"
 	CmdGetData        = "getdata"
 	CmdNotFound       = "notfound"
 	CmdBlock          = "block"
+	CmdLightBlock     = "lightblock"
 	CmdTx             = "tx"
 	CmdGetHeaders     = "getheaders"
 	CmdHeaders        = "headers"
@@ -90,8 +92,14 @@ func makeEmptyMessage(command string) (Message, error) {
 	case CmdBlock:
 		msg = &MsgBlock{}
 
+	case CmdLightBlock:
+		msg = &MsgLightBlock{}
+
 	case CmdInv:
 		msg = &MsgInv{}
+
+	case CmdGetMissedTxs:
+		msg = &MsgGetMissedTxs{}
 
 	case CmdGetData:
 		msg = &MsgGetData{}