+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Hcash Daemon   
 
+## [2017-11-13]  
++ **Updated**  
+  - done all tests cases under package `hcashec/edward`   
+
+## [2017-11-10]  
++ **Updated**  
+  - test cases under package `hcashec/edward`  
+
 ## [2017-10-25]  
 + **Fixed**   
   - all test cases under package `blockchain`